:recycle: #4 Mock 객체들을 반환해주는 StubDataUtil 생성. 여기서 Mock 객체들을 받아서 Test 메소드에서 사용하는 구조.
