+++ src/test/java/com/seollem/server/restdocs/controller/book/Calender.java
@@ -1,0 +1,57 @@
+package com.seollem.server.restdocs.controller.book;
+
+import static org.mockito.BDDMockito.given;
+
+import com.google.gson.Gson;
+import com.seollem.server.book.BookController;
+import com.seollem.server.book.BookMapper;
+import com.seollem.server.book.BookService;
+import com.seollem.server.member.MemberService;
+import com.seollem.server.memo.MemoMapper;
+import com.seollem.server.memo.MemoService;
+import com.seollem.server.memolikes.MemoLikesService;
+import com.seollem.server.restdocs.util.WebMvcTestSetUpUtil;
+import com.seollem.server.util.GetEmailFromHeaderTokenUtil;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mockito;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+
+@ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
+@WebMvcTest(BookController.class)
+public class Calender extends WebMvcTestSetUpUtil {
+
+  @MockBean
+  private GetEmailFromHeaderTokenUtil getEmailFromHeaderTokenUtil;
+  @MockBean
+  private MemberService memberService;
+  @MockBean
+  private BookService bookService;
+  @MockBean
+  private BookMapper bookMapper;
+  @MockBean
+  private MemoService memoService;
+  @MockBean
+  private MemoMapper memoMapper;
+  @MockBean
+  private MemoLikesService memoLikesService;
+
+  @Autowired
+  private Gson gson;
+
+  @Test
+  public void calenderTest() throws Exception {
+    //given
+    given(getEmailFromHeaderTokenUtil.getEmailFromHeaderToken(Mockito.anyMap())).willReturn(
+        "starrypro@gmail.com");
+
+    //when
+
+    //then
+  }
+}
